Modified shared_memory to save name.

+++ foundation/ipc/shared_memory.cpp
@@ -14,65 +14,72 @@ namespace
 {
 	using namespace ooe;
 
-	descriptor memory_open( const std::string& name, ipc::shared_memory::type mode, up_t size )
-	{
-		s32 flags = mode == ipc::shared_memory::open ? O_RDWR : O_RDWR | O_CREAT | O_EXCL;
-		descriptor desc( shm_open( name.c_str(), flags, 0600 ) );
-
-		if ( mode == ipc::shared_memory::create )
-			desc.resize( size );
-
-		return desc;
-	}
-}
-
-namespace ooe
-{
-//--- ipc::shared_memory_id ----------------------------------------------------
-	ipc::shared_memory_id::shared_memory_id( const std::string& name_ )
-		: name( name_ )
+	void shm_remove( const std::string& name )
 	{
 		if ( shm_unlink( name.c_str() ) && errno != OOE_NOEXIST )
 			throw error::runtime( "ipc::shared_memory: " ) <<
 				"Unable to unlink shared memory \"" << name << "\": " << error::number( errno );
 	}
 
-	ipc::shared_memory_id::~shared_memory_id( void )
+	descriptor shm_create( const std::string& name, bool create, up_t size )
 	{
-		if ( shm_unlink( name.c_str() ) && errno != OOE_NOEXIST )
+		if ( create )
+			shm_remove( name );
+
+		s32 flags = create ? O_RDWR | O_CREAT | O_EXCL : O_RDWR;
+		descriptor desc( shm_open( name.c_str(), flags, 0600 ) );
+
+		if ( create )
+			desc.resize( size );
+
+		return desc;
+	}
+
+	void shm_destroy( const std::string& name, bool unlinkable )
+	{
+		if ( unlinkable && shm_unlink( name.c_str() ) && errno != OOE_NOEXIST )
 			OOE_WARNING( "ipc::shared_memory",
 				"Unable to unlink shared memory \"" << name << "\": " << error::number( errno ) );
 	}
+}
 
+namespace ooe
+{
 //--- ipc::shared_memory_base --------------------------------------------------
-	ipc::shared_memory_base::shared_memory_base( const std::string& name_, bool open )
-		: id( open ? 0 : new shared_memory_id( name_ ) )
+	ipc::shared_memory_base::shared_memory_base( const std::string& shm_name, bool unlinkable_ )
+		: name_( shm_name ), unlinkable( unlinkable_ )
 	{
 	}
 
 	ipc::shared_memory_base::~shared_memory_base( void )
 	{
+		shm_destroy( name_, unlinkable );
 	}
 
 	std::string ipc::shared_memory_base::name( void ) const
 	{
-		return id ? id->name : std::string();
+		return name_;
 	}
 
 	void ipc::shared_memory_base::unlink( void )
 	{
-		delete id.release();
+		if ( unlinkable )
+		{
+			shm_remove( name_ );
+			name_ = std::string();
+			unlinkable = false;
+		}
 	}
 
 //--- ipc::shared_memory -------------------------------------------------------
-	ipc::shared_memory::shared_memory( const std::string& name_, type mode, up_t size_ )
-		: shared_memory_base( name_, mode == open ),
-		memory( memory_open( name_, mode, size_ ), memory::read_write )
+	ipc::shared_memory::shared_memory( const std::string& shm_name, type mode, up_t size_ )
+		: shared_memory_base( shm_name, mode == create ),
+		memory( shm_create( shm_name, mode == create, size_ ), memory::read_write )
 	{
 	}
 
-	ipc::shared_memory::shared_memory( const ooe::descriptor& desc )
-		: shared_memory_base( std::string(), true ), memory( desc, memory::read_write )
+	ipc::shared_memory::shared_memory( const std::string& shm_name, const ooe::descriptor& desc )
+		: shared_memory_base( shm_name, false ), memory( desc, memory::read_write )
 	{
 	}
 
@@ -81,12 +88,12 @@ namespace ooe
 	}
 
 //--- ipc::locked_memory -------------------------------------------------------
-	ipc::locked_memory::locked_memory( const std::string& name_, type mode, up_t size_ )
-		: shared_memory( name_, mode, size_ )
+	ipc::locked_memory::locked_memory( const std::string& shm_name, type mode, up_t size_ )
+		: shared_memory( shm_name, mode, size_ )
 	{
 		if ( mlock( get(), size() ) )
 			throw error::runtime( "ipc::locked_memory: " ) <<
-				"Unable to lock shared memory \"" << name_ << "\": " << error::number( errno );
+				"Unable to lock shared memory \"" << shm_name << "\": " << error::number( errno );
 	}
 
 	ipc::locked_memory::~locked_memory( void )
